moviesSaga changed to listSaga and moved to features
