+++ src/rootSaga.js
@@ -1,6 +1,6 @@
 import { all } from "redux-saga/effects";
 import { watchSetMovie } from "./features/movies/MoviePage/movieSaga";
-import { watchSetActivePage } from "./features/movies/MoviesPage/moviesSaga"
+import { watchSetActivePage } from "./features/listSaga"
 
 export default function* rootSaga() {
     yield all([
